ai-service/requirements.txt
fastapi
uvicorn
scikit-learn
pdfminer.six
sentence-transformers
python-multipart
google-generativeai
python-dotenv
g4f
curl_cffi